ref: rm `TaskStorage` trait and rename `FileTaskStorage` to `TaskStorage`
Having a trait to implement a in memory storage for tests was a good
idea, but it was forcing us to make a lot of thing mutable, which was
not something we want.

Instead we remove the trait and in-memory implementation to only use the
file based one with tempfile.