feat: allow the use of `#inbox` and `#backlog` to influence the sorting
The `#inbox` hashtag will add a couple point to the urgency, because it
means those tasks are what you want to do today or at least it's the
next bunch of tasks you want to tackle.

The `#backlog` hashtag remove 30 point to the urgency, because even if
they're assigned to a project and / or a context, they're not urgent at
all and shouldn't be in the scope.